+++ app/src/main/java/ch/threema/app/webclient/services/instance/message/receiver/ModifyDistributionListHandler.java
@@ -86,7 +86,7 @@ public class ModifyDistributionListHandler extends MessageReceiver {
 			this.failed(temporaryId, Protocol.ERROR_BAD_REQUEST);
 			return;
 		}
-		final int distributionListId = Integer.parseInt(args.get(Receiver.ID).asStringValue().toString());
+		final long distributionListId = Long.parseLong(args.get(Receiver.ID).asStringValue().toString());
 
 		// Get distribution list
 		final DistributionListModel distributionListModel = this.distributionListService.getById(distributionListId);
